# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renault Parts Alexandria - API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  phone: string;
  email: string;
  password: string;
  /** @nullable */
  carModel?: string | null;
  /** @nullable */
  carYear?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  area?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  email: string;
  /** @nullable */
  carModel?: string | null;
  /** @nullable */
  carYear?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  area?: string | null;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserBody {
  name?: string;
  phone?: string;
  /** @nullable */
  carModel?: string | null;
  /** @nullable */
  carYear?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  area?: string | null;
}

export interface Part {
  id: number;
  name: string;
  /** @nullable */
  oemCode?: string | null;
  type: string;
  /** @nullable */
  priceOriginal?: number | null;
  /** @nullable */
  priceTurkish?: number | null;
  /** @nullable */
  priceChinese?: number | null;
  /** @nullable */
  compatibleModels?: string | null;
  /** @nullable */
  supplier?: string | null;
}

export interface Package {
  id: number;
  name: string;
  slug: string;
  description: string;
  kmService: number;
  basePrice: number;
  sellPrice: number;
  warrantyMonths: number;
  parts: Part[];
  createdAt: string;
}

export interface Workshop {
  id: number;
  name: string;
  area: string;
  address: string;
  phone: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  rating?: number | null;
  partnershipStatus: string;
}

export interface CreateOrderBody {
  packageId: number;
  /** @nullable */
  workshopId?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryArea?: string | null;
  paymentMethod: string;
  carModel: string;
  carYear: number;
  /** @nullable */
  notes?: string | null;
}

export interface Order {
  id: number;
  userId: number;
  packageId: number;
  /** @nullable */
  workshopId?: number | null;
  status: string;
  paymentMethod: string;
  paymentStatus: string;
  total: number;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryArea?: string | null;
  carModel: string;
  carYear: number;
  /** @nullable */
  notes?: string | null;
  package: Package;
  workshop?: Workshop | null;
  createdAt: string;
}

export interface CreateReviewBody {
  orderId: number;
  /** @nullable */
  workshopId?: number | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface Review {
  id: number;
  orderId: number;
  userId: number;
  /** @nullable */
  workshopId?: number | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ChatMessageBody {
  message: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  carModel?: string | null;
  /** @nullable */
  carYear?: number | null;
  /** @nullable */
  mileage?: number | null;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
  /** @nullable */
  suggestedPackageSlug?: string | null;
  /** @nullable */
  suggestedPackageName?: string | null;
}

export type ListPartsParams = {
  packageId?: number;
};

export type ListWorkshopsParams = {
  area?: string;
};