# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Renault Parts Alexandria - API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: packages
    description: Package operations
  - name: parts
    description: Parts operations
  - name: workshops
    description: Workshop operations
  - name: orders
    description: Order operations
  - name: reviews
    description: Review operations
  - name: chat
    description: AI Chat operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current logged in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"

  /packages/{slug}:
    get:
      operationId: getPackageBySlug
      tags: [packages]
      summary: Get package by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List all parts
      parameters:
        - name: packageId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Part"

  /workshops:
    get:
      operationId: listWorkshops
      tags: [workshops]
      summary: List all workshops
      parameters:
        - name: area
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of workshops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workshop"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send message to AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        carModel:
          type: ["string", "null"]
        carYear:
          type: ["integer", "null"]
        address:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
      required:
        - name
        - phone
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        carModel:
          type: ["string", "null"]
        carYear:
          type: ["integer", "null"]
        address:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - email
        - role
        - createdAt

    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        carModel:
          type: ["string", "null"]
        carYear:
          type: ["integer", "null"]
        address:
          type: ["string", "null"]
        area:
          type: ["string", "null"]

    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        kmService:
          type: integer
        basePrice:
          type: number
        sellPrice:
          type: number
        warrantyMonths:
          type: integer
        parts:
          type: array
          items:
            $ref: "#/components/schemas/Part"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - kmService
        - basePrice
        - sellPrice
        - warrantyMonths
        - parts
        - createdAt

    Part:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        oemCode:
          type: ["string", "null"]
        type:
          type: string
        priceOriginal:
          type: ["number", "null"]
        priceTurkish:
          type: ["number", "null"]
        priceChinese:
          type: ["number", "null"]
        compatibleModels:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
      required:
        - id
        - name
        - type

    Workshop:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        area:
          type: string
        address:
          type: string
        phone:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        partnershipStatus:
          type: string
      required:
        - id
        - name
        - area
        - address
        - phone
        - partnershipStatus

    CreateOrderBody:
      type: object
      properties:
        packageId:
          type: integer
        workshopId:
          type: ["integer", "null"]
        deliveryAddress:
          type: ["string", "null"]
        deliveryArea:
          type: ["string", "null"]
        paymentMethod:
          type: string
        carModel:
          type: string
        carYear:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - packageId
        - paymentMethod
        - carModel
        - carYear

    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
        workshopId:
          type: ["integer", "null"]
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        total:
          type: number
        deliveryAddress:
          type: ["string", "null"]
        deliveryArea:
          type: ["string", "null"]
        carModel:
          type: string
        carYear:
          type: integer
        notes:
          type: ["string", "null"]
        package:
          $ref: "#/components/schemas/Package"
        workshop:
          oneOf:
            - $ref: "#/components/schemas/Workshop"
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - packageId
        - status
        - paymentMethod
        - paymentStatus
        - total
        - carModel
        - carYear
        - package
        - createdAt

    CreateReviewBody:
      type: object
      properties:
        orderId:
          type: integer
        workshopId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required:
        - orderId
        - rating

    Review:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        userId:
          type: integer
        workshopId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - userId
        - rating
        - createdAt

    ChatMessageBody:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: ["string", "null"]
        carModel:
          type: ["string", "null"]
        carYear:
          type: ["integer", "null"]
        mileage:
          type: ["integer", "null"]
      required:
        - message

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        sessionId:
          type: string
        suggestedPackageSlug:
          type: ["string", "null"]
        suggestedPackageName:
          type: ["string", "null"]
      required:
        - reply
        - sessionId