# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renault Parts Alexandria - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  packageId: number;
  /** @nullable */
  workshopId?: number | null;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryArea?: string | null;
  paymentMethod: string;
  carModel: string;
  carYear: number;
  /** @nullable */
  notes?: string | null;
}