# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renault Parts Alexandria - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Package } from "./package";
import type { Workshop } from "./workshop";

export interface Order {
  id: number;
  userId: number;
  packageId: number;
  /** @nullable */
  workshopId?: number | null;
  status: string;
  paymentMethod: string;
  paymentStatus: string;
  total: number;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryArea?: string | null;
  carModel: string;
  carYear: number;
  /** @nullable */
  notes?: string | null;
  package: Package;
  workshop?: Workshop | null;
  createdAt: Date;
}