# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Renault Parts Alexandria - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterBody {
  name: string;
  phone: string;
  email: string;
  password: string;
  /** @nullable */
  carModel?: string | null;
  /** @nullable */
  carYear?: number | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  area?: string | null;
}